backend/requirements.txt
# Flask and Extensions
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.43

# Security and Authentication
Werkzeug==3.1.3
bcrypt==4.2.1
PyJWT==2.10.1
Flask-Limiter==3.8.0
bleach==6.2.0

# Environment and Configuration
python-dotenv==1.1.1

# HTTP Requests
requests==2.32.5

# AI/ML Libraries
transformers==4.56.0
torch==2.8.0
numpy==2.3.2

# Utilities
click==8.2.1
colorama==0.4.6
markupsafe==3.0.2
Jinja2==3.1.6

# File Handling
Pillow==11.1.0

# Development Dependencies
flake8==7.1.1
pytest==8.3.4
pytest-flask==1.3.0

# Production Dependencies
gunicorn==23.0.0